requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiosqlite>=0.20.0
httpx>=0.27.0
python-multipart>=0.0.12
websockets>=13.0
llama-cpp-python>=0.3.0
huggingface-hub>=0.24.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
